Station and Position Trigger sparse array improvements
With a high mover count (>100) and a high station count (>100) it can take a significant amount of test all movers at all stations as shown below. Profiler shows 200 stations and 170 movers takes ~3ms per station totaling 600ms of Station.Cyclic() each scan.
A common use case with stations, especially with mixed and matched station groupings is to number stations by 10s thus the station array would be 1..100, but only use some of the stations reflecting different 2-up, 4-up and other arrangements.
In this scenario these stations may actually be used in the array 10, 11, 20, 21, 22, 23, 30, 31...
By setting an enabled flag, only stations that have been assigned a position will be processed allowing for a sparse station array
without a performance penalty.
Enabling is transparent to the user and accomplished by setting .Position which already happens as part of the initializing process
The same is applied to position triggers.