Make station.cyclic call station.MoverInPosition less
By storing the mover in position bit at the top of the .cyclic routine, we reduce the number of times mover in position is calculated for the the cyclic routine from 6 to 1. With 170 movers this reduced the station.cyclic scan time from 4.0ms to 1.6ms.
The property .MoverInPosition is still calculated on every call to it from outside the cyclic routine.